service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech2text-extension-service"
version = "1.2.0"
description = "D-Bus service providing speech-to-text functionality for GNOME Shell"
readme = "README.md"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
authors = [
    {name = "Kaveh Tehrani", email = "codemonkey13x@gmail.com"}
]
maintainers = [
    {name = "Kaveh Tehrani", email = "codemonkey13x@gmail.com"}
]
keywords = [
    "gnome",
    "speech-to-text",
    "dbus",
    "voice-recognition",
    "accessibility"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment :: Gnome",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "dbus-next>=0.2.3",
    "openai-whisper>=20231117",
    "torch>=1.13.0",
    "torchaudio>=0.13.0",
]

[project.optional-dependencies]
# Optional HTTP server (remote GPU box)
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

dev = [
    "build",
    "twine",
    "pytest",
]

[project.urls]
Homepage = "https://github.com/kavehtehrani/speech2text-extension"
Repository = "https://github.com/kavehtehrani/speech2text-extension"
Issues = "https://github.com/kavehtehrani/speech2text-extension/issues"
Documentation = "https://github.com/kavehtehrani/speech2text-extension/blob/main/README.md"

[project.scripts]
speech2text-extension-service = "gnome_speech2text_service.cli:main"
speech2text-extension-remote-server = "gnome_speech2text_service.remote_server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gnome_speech2text_service = ["../data/*"]
